+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # cuGraph 0.7.0 (Date TBD)
 
 ## New Features
-
+- PR #195 Added Graph.get_two_hop_neighbors() method
+- PR #195 Updated Jaccard and Weighted Jaccard to accept lists of vertex pairs to compute for
 ...
 
 ## Improvements